finish with second task - init webpack, make modules
